% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_uncertainty}
\alias{calc_uncertainty}
\title{Calculate poll uncertainty}
\usage{
calc_uncertainty(p, n, z = 1.96)
}
\arguments{
\item{p}{share of poll}

\item{n}{number of respondents}

\item{z}{the z-score}
}
\value{
a number
}
\description{
Uses a formula to calculate poll uncertainty
}
\examples{
calc_uncertainty(.2, 1000)
}
