% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{popAddTrend}
\alias{popAddTrend}
\title{Add Trend to Polls}
\usage{
popAddTrend(
  data,
  name = NULL,
  type = "kalman",
  args = list(),
  interpolations = list()
)
}
\arguments{
\item{data}{A \code{popPolls} object.}

\item{name}{Name of the trend. Defaults to a name built from \code{type} and the
applied interpolations.}

\item{type}{Trend function, see details.}

\item{args}{Arguments passed on to the trend function.}

\item{interpolations}{Named list of interpolations that should be applied to
the trend, see details.}
}
\value{
The \code{popPolls} object with the trend added to \verb{$trends}.
}
\description{
Calculates a poll aggregation and stores it in the \code{popPolls} object, where
it is picked up by \code{\link[=plot.popPolls]{plot.popPolls()}}.
}
\details{
Available trend functions are:

\describe{
\item{\code{kalman}}{Kalman filter, arguments: \code{sd = 0.003}.}
\item{\code{kalmanKFAS}}{Kalman filter based on the \pkg{KFAS} package,
arguments: \code{sd = 0.003}, \code{smoothing = TRUE}.}
\item{\code{weightedMeanLastDays}}{Linearly weighted rolling mean, arguments:
\code{days = 30}, \code{maxObs = Inf}.}
\item{\code{ident}}{Plain mean of all polls published on the same day, no
arguments.}
}

Available interpolations are:

\describe{
\item{\code{lastInterpolation}}{Carries the last value forward, no arguments.}
\item{\code{linearInterpolation}}{Linear interpolation, no arguments.}
\item{\code{bernoulliConvInterpolation}}{Binomial smoothing, arguments:
\code{n = 20}, \code{k = 6}.}
}
}
\examples{
\dontrun{
de = popRead('DE-parliament')
de = popAddTrend(de, name = 'Kalman 0.003', type = 'kalman', args = list(sd = 0.003))
de = popAddTrend(de, name = 'Kalman Raw', type = 'kalman', args = list(sd = 0.003),
                 interpolations = list('lastInterpolation' = list()))
plot(de)
}
}
