% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pp_add_weights}
\alias{pp_add_weights}
\title{Adds weights to the polls}
\usage{
pp_add_weights(checked_polls, pollster_rating = pp_calc_pollster_rating())
}
\arguments{
\item{checked_polls}{the output of pp_check_raw_polls()}

\item{pollster_rating}{the pollster rating. Defaults to the
pp_calc_pollster_rating() function.}
}
\value{
tibble
}
\description{
Calculates and adds weights to the polls
}
