% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pp_auto_poll_of_poll}
\alias{pp_auto_poll_of_poll}
\title{Driverless poll of polls}
\usage{
pp_auto_poll_of_poll()
}
\value{
tibble
}
\description{
Combines all functions in to one driverless poll of polls function
}
