% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pp_calc_poll}
\alias{pp_calc_poll}
\title{Calculate the final poll of polls}
\usage{
pp_calc_poll(polls_with_wt)
}
\arguments{
\item{polls_with_wt}{the output of pp_add_weights()}
}
\value{
tibble
}
\description{
Calculate the final poll of polls
}
