% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pp_calc_pollster_rating}
\alias{pp_calc_pollster_rating}
\title{Calculates the pollster rating}
\usage{
pp_calc_pollster_rating()
}
\value{
tibble
}
\description{
Based on <https://fivethirtyeight.com/features/the-polls-are-all-right/>
article on how to calculate a pollster rating.
}
