% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pp_check_raw_polls}
\alias{pp_check_raw_polls}
\title{Checks and recalculates the polls}
\usage{
pp_check_raw_polls(raw_polls, silent = FALSE)
}
\arguments{
\item{raw_polls}{the output of pp_get_raw_polls()}

\item{silent}{should the function run silent?}
}
\value{
tibble
}
\description{
Checks the polls and recalculates them if they do not sum to 100.
}
