% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polls.R
\name{pp_get_raw_polls}
\alias{pp_get_raw_polls}
\title{Download raw polls}
\usage{
pp_get_raw_polls(days = 28, source = "berlingske")
}
\arguments{
\item{days}{keeps only polls that are equal to or younger than the days
parameter.}

\item{source}{where to get the polls? Defaults to Berlingske}
}
\value{
tibble
}
\description{
Downloads raw polls from either Berlingske Barometer or Erik Gahner.
Downloads all but saves only the latest xx days as set by the day parameter.
}
\examples{
polls <- pp_get_raw_polls_berlingske()
}
