% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polls.R
\name{pp_get_raw_polls_gahner}
\alias{pp_get_raw_polls_gahner}
\title{Download raw polls from Gahner}
\usage{
pp_get_raw_polls_gahner(days = 28)
}
\arguments{
\item{days}{keeps only polls that are equal to or younger than the days
parameter.}
}
\value{
tibble
}
\description{
Downloads raw polls from Erik Gahners github repo. Downloads all but saves only
the latest xx days as set by the day parameter
}
\examples{
polls <- pp_get_raw_polls_gahner()
}
