% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_of_polls.R
\name{pp_poll_of_polls_summariser}
\alias{pp_poll_of_polls_summariser}
\title{Poll of Polls by summarising}
\usage{
pp_poll_of_polls_summariser(polls, collapse_firm = TRUE)
}
\arguments{
\item{polls}{a data set of polls}

\item{collapse_firm}{collapse multiple polls from one firm into one poll that
is the mean of that firms polls?}
}
\value{
a tibble
}
\description{
This method calculates the number of respondents for each party in each poll
by taking the total number of respondents for the poll  and multiplying it
with the share of each party.
}
\details{
Then it pools all the respondents for each party together and all the
respondents from all the polls together to bassically create one massive poll.

This poll is then used to calculate the percentages and uncertainties anew.
}
\examples{
polls <- pp_get_raw_polls()
pp_poll_of_polls_summariser(polls)
}
