% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{round_preserve_sum}
\alias{round_preserve_sum}
\title{Rounds but preserves sum}
\usage{
round_preserve_sum(x, digits = 0)
}
\arguments{
\item{x}{numbers to round}

\item{digits}{digits to keep}
}
\value{
a vector
}
\description{
Rounds but preserves sum
}
\examples{
x <- rep(10 / 3, 3)
round_preserve_sum(x)
}
