% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{plot.popPolls}
\alias{plot.popPolls}
\title{Plot polls}
\usage{
\method{plot}{popPolls}(data, xlim = NULL)
}
\arguments{
\item{data}{Polls Data}

\item{xlim}{}
}
\description{
Plot polls
}
\examples{
t = popRead('DE')
plot(t)
plot(t, xlim=as.Date(c('2018-01-01', '2018-05-01')))
}
