% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{plot.popPolls}
\alias{plot.popPolls}
\title{Plot polls}
\usage{
\method{plot}{popPolls}(x, ...)
}
\arguments{
\item{x}{Polls Data}

\item{...}{xlim: Limit date range}
}
\description{
Plot polls
}
\examples{
t = popRead('DE-parliament')
plot(t)
plot(t, xlim=as.Date(c('2018-01-01', '2018-05-01')))
}
