% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{popAddTrend}
\alias{popAddTrend}
\title{Add Trend to Polls}
\usage{
popAddTrend(data, name = NULL, type = "kalman", args = list(),
  interpolations = list())
}
\arguments{
\item{data}{polls object}

\item{name}{trend name}

\item{type}{trend function name}

\item{args}{trend function arguments}

\item{interpolations}{list of interpolations that should be applied to the trend}
}
\value{
polls object
}
\description{
Add Trend to Polls
}
\details{
Available trend function names are:

\code{kalman} - arguments: sd = 0.03

\code{weightedMeanLastDays} - arguments: days = 30, maxObs = Inf


Available interpolations are:

\code{lastInterpolation} - no arguments.

\code{linearInterpolation} - no arguments.

\code{bernoulliConvInterpolation} - arguments: n = 20, k = 6
}
\examples{
t = popRead('DE')
t = popAddTrend(t, name='Kalman 0.05', type='kalman', args=list(sd = 0.05), )
t = popAddTrend(t, name='Kalman raw', type='kalman', args=list(sd = 0.02), interpolations=list('lastInterpolation' = list()))
plot(t)
}
