% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{popCreate}
\alias{popCreate}
\title{Create pop object}
\usage{
popCreate(polls = data.table(), options = list(measure = "p"),
  parties = data.table(), trends = list(), name = NULL,
  elections = data.table())
}
\arguments{
\item{polls}{polls}

\item{options}{options}

\item{parties}{parties}

\item{trends}{trends}

\item{name}{name}

\item{elections}{elections}
}
\value{
pop object
}
\description{
Create pop object
}
