% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{popRead}
\alias{popRead}
\title{Read Poll Data}
\usage{
popRead(code, load = c("polls", "elections"))
}
\arguments{
\item{code}{Code of the poll data.}

\item{load}{what to load.}
}
\value{
poll data
}
\description{
Read Poll Data
}
\examples{
popRead('DE-parliament')
}
