% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pp_get_raw_polls}
\alias{pp_get_raw_polls}
\title{Download raw polls}
\usage{
pp_get_raw_polls(days = 28)
}
\arguments{
\item{days}{keeps only polls that are equal to or younger than the days
parameter.}
}
\value{
tibble
}
\description{
Downloads raw polls from Erik Gahners github repo. Downloads all but only
save the latest xx days as set by the day parameter
}
