% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{print.popPolls}
\alias{print.popPolls}
\title{Print popPolls object}
\usage{
\method{print}{popPolls}(x, ...)
}
\arguments{
\item{x}{popPolls object}

\item{...}{ignored}
}
\description{
Print popPolls object
}
